+++ src/main/kotlin/blackjack/domain/participant/Dealer.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.player
+package blackjack.domain.participant
 
 import blackjack.domain.card.Cards
