+++ src/main/kotlin/blackjack/domain/participant/Participants.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.player
+package blackjack.domain.participant
 
 class Participants(val participants: List<Participant>) {
     init {
