+++ src/test/kotlin/blackjack/domain/participant/DealerTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.player
+package blackjack.domain.participant
 import io.kotest.core.spec.style.StringSpec
 
 class DealerTest : StringSpec({
