+++ src/test/kotlin/blackjack/domain/participant/PlayerTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.player
+package blackjack.domain.participant
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.CardNumber
